refactor(offer-screen): выносит логику компонента в хук
